Uses long for download byte counters

+++ project/SPTarkov.Core/Mods/DownloadTask.cs
@@ -6,7 +6,7 @@ public class DownloadTask : IModTask
 {
     public required ForgeBase ForgeMod { get; init; }
     public required ForgeModVersion Version { get; init; }
-    public float TotalToDownload { get; set; }
+    public long TotalToDownload { get; set; }
     public float Progress { get; set; }
     public required CancellationTokenSource CancellationTokenSource { get; set; }
     public bool Complete { get; set; }
