+++ project/SPTarkov.Core/Mods/InstallTask.cs
@@ -6,7 +6,7 @@ public class InstallTask : IModTask
 {
     public required ConfigMod ForgeMod { get; init; }
     public required CancellationTokenSource CancellationTokenSource { get; set; }
-    public float TotalToDownload { get; set; }
+    public long TotalToDownload { get; set; }
     public float Progress { get; set; }
     public bool Complete { get; set; }
     public Exception? Error { get; set; }
